Conflicts:
	.gitignore
	Android/app/app.iml
	Android/app/build.gradle
	Android/app/src/main/jni/Android.mk
	Android/lib/src/main/jni/Android.mk
	Android/lib/wkcocos.iml
	Classes/App/AppDelegate.cpp
	Classes/App/CMakeLists.txt
	Classes/App/ErrorUI.cpp
	Classes/App/HelloWorldScene.cpp
	Classes/App/MyPlayer.cpp
	Classes/App/SavingScene.cpp
	Classes/Lib/CMakeLists.txt
	Include/WkCocos/Player.h
	Include/WkCocosApp/GameLogic.h
	depends/cocos2d-x

+++ Include/WkCocos/Download/Systems/DLCchecking.h
@@ -25,10 +25,10 @@ namespace WkCocos
 
 				void update(entityx::ptr<entityx::EntityManager> es, entityx::ptr<entityx::EventManager> events, double dt) override;
 
-				void setConnectionTimeout(unsigned int timeout);
+				void setConnectionTimeout(unsigned long timeout);
 
 				void *_curl;
-				unsigned int _connectionTimeout;
+				unsigned long _connectionTimeout;
 
 				std::string m_manifest;
